Wait for 2Secs to get next period start

+++ admanager_mpd.h
@@ -373,7 +373,7 @@ public:
 	bool                                           mExitFulfillAdLoop;    /**< Flag to exit the Ad fulfillment loop */
 	std::mutex                                     mAdPlacementMtx;       /**< Mutex protecting Ad placement */
 	std::condition_variable                        mAdPlacementCV;        /**< Condition variable for Ad placement */
-	int                                            mWaitForManifestUpdate;/**< Counter for waiting manifest updates */
+	int64_t                                        mWaitForManifestUpdate;/**< mS time for waiting manifest updates */
 	/**
 	 * @fn PrivateCDAIObjectMPD
 	 *
